enabled 32bits build

+++ cctools-repo/build.sh
@@ -13,27 +13,31 @@ WRKDIR=${WRKDIR}-pie
 # -------------------------------
 #
 
-SDKDIR=/home/sash/Android/android-sdk
-NDKDIR=/home/sash/Android/android-ndk
-NDKSRC=/home/sash/Android/ndk-src
+SDKDIR=/home/sash/Android/sdk
+NDKDIR=/mnt/sash/android-ndk-r12b
+NDKSRC=/mnt/sash/ndk-r12b-src
 
 for d in binutils gcc gmp mpc mpfr cloog isl ppl llvm-3.6; do
-    ln -sf ${NDKSRC}/${d} src/
+    if [ -e ${NDKSRC}/${d} -a ! -e src/${d} ]; then
+	ln -sf ${NDKSRC}/${d} src/
+    else
+	echo "Skipping $d"
+    fi
 done
 
-export PATH=~/bin/CodeSourcery/bin:$PATH
+export PATH=/mnt/sash/CodeSourcery/bin:$PATH
 
-#./build-shell-utils.sh ${PWD}/src arm-linux-androideabi  ${WRKDIR}/arm-repo $NDKDIR $SDKDIR      || exit 1
+./build-shell-utils.sh ${PWD}/src arm-linux-androideabi  ${WRKDIR}/arm-repo $NDKDIR $SDKDIR      || exit 1
 
-#./build-shell-utils.sh ${PWD}/src mipsel-linux-android   ${WRKDIR}/mips-repo $NDKDIR $SDKDIR     || exit 1
+./build-shell-utils.sh ${PWD}/src mipsel-linux-android   ${WRKDIR}/mips-repo $NDKDIR $SDKDIR     || exit 1
 
-#./build-shell-utils.sh ${PWD}/src i686-linux-android     ${WRKDIR}/i686-repo $NDKDIR $SDKDIR     || exit 1
+./build-shell-utils.sh ${PWD}/src i686-linux-android     ${WRKDIR}/i686-repo $NDKDIR $SDKDIR     || exit 1
 
-./build-shell-utils.sh ${PWD}/src aarch64-linux-android  ${WRKDIR}/arm64-repo $NDKDIR $SDKDIR  || exit 1
+#./build-shell-utils.sh ${PWD}/src aarch64-linux-android  ${WRKDIR}/arm64-repo $NDKDIR $SDKDIR  || exit 1
 
 #./build-shell-utils.sh ${PWD}/src mips64el-linux-android ${WRKDIR}/mips64-repo $NDKDIR $SDKDIR || exit 1
 
-./build-shell-utils.sh ${PWD}/src x86_64-linux-android   ${WRKDIR}/x86_64-repo $NDKDIR $SDKDIR   || exit 1
+#./build-shell-utils.sh ${PWD}/src x86_64-linux-android   ${WRKDIR}/x86_64-repo $NDKDIR $SDKDIR   || exit 1
 
 test -e ${WRKDIR}/repo/armeabi-v7a || ln -sf armeabi ${WRKDIR}/repo/armeabi-v7a
 test -e ${WRKDIR}/repo/mips-r2     || ln -sf mips    ${WRKDIR}/repo/mips-r2
